// task-management/src/app/_core/api/rolePermission/role-permission.ts
export interface rolePermission {
  pagingData: pagingData;
}

export interface pagingData {
  content: [rolePermissionContent];
  pageable: pageable;
  totalElements: number;
  totalPages?: number;
  size?: number;
}

export interface rolePermissionContent {
  id?: number;
  permissionId?: number;
  roleId?: number;
  isChecked?: Boolean;
}

export interface pageable {
  pageNumber: number;
  pageSize: number;
}